+++ chem-prac-python/PeriodicPractor[Origin].py
